requirements.txt
# Core dependencies
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
sqlalchemy>=2.0.0
pyodbc>=4.0.39

# Simulation and calculations
simpy>=4.0.1
scipy>=1.11.0

# Visualization and reporting
plotly>=5.17.0
openpyxl>=3.1.0

# Configuration and utilities
python-dotenv>=1.0.0

# Security dependencies
bcrypt>=4.0.1
cryptography>=41.0.0
pyjwt>=2.8.0

# Development and testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Logging and monitoring
logging>=0.4.9.6